pyproject.toml
[tool.poetry]
name = "collector"
version = "0.2.2"
description = "Collector bundle with a sample Scraper that is forkable and meant to be forked"
authors = ["Benedikt Schäfer", "Noah Lokocz", "Liv Märtens"]
readme = "README.md"
license = "NoLicense"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.22.0"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
pydantic = "^2.9.2"
aiohttp = "^3.11.10"
redis = "^5.2.1"
litellm = "^1.67"
pytest-asyncio = "^0.25.2"
pypdf = "^5.1.0"
openapi-client = {path = "oapicode"}
jsondiff = "^2.2.1"
kreuzberg = {extras = ["easyocr"], version = "^3.18.0"}
toml = "^0.10.2"
dotenv = "^0.9.9"
jsonschema = "^4.25.1"
lazy-imports = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
